+++ generators/app/promptConfig.js
@@ -9,12 +9,6 @@ const promptConfig = () => {
             {
                 type: 'confirm',
                 default: false,
-                name: 'githook_postcommit',
-                message: "Do you want to install post-commit hook to create change files for each (non-chore) commit?"
-            },
-            {
-                type: 'confirm',
-                default: false,
                 name: 'githook_prepush',
                 message: "Do you want to install pre-push hook to ensure change files exist? 'rush change -v' will be invoked on every 'git push'."
             }