+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log - generator-rush-conventionalcommits
 
+## 1.4.0
+
+- `rush changefiles` command and `post-commit` git hook removed
+- `rush whatchanged` command detects existing change files and parses commits made after the last change file's creation date (based on the date in the file name)
+
 ## 1.3.0
 
 ### Features